requirements.txt
numpy==1.16.4
pandas==0.24.2
Flask_Cors==3.0.8
Flask_Bcrypt==0.7.1
Flask==1.1.1
Werkzeug==0.15.5
flask_restplus==0.12.1
Flask_SQLAlchemy==2.4.0
PyMySQL==0.9.3
cx_Oracle==7.2.3
python-dotenv==0.10.3
scikit_learn==0.21.3